vdr: fix old config import.

+++ src/paks/vdr/update.sh
@@ -29,7 +29,7 @@ OLD_DIR="/opt/vdr/etc"
 NEW_DIR="/etc/vdr"
 
 for conf in ${OLD_DIR}/*.conf; do
-	[ -f "${conf}" ] && continue
+	[ -f "${conf}" ] || continue
 
 	conf="$(basename ${conf})"
 	mv -b "${OLD_DIR}/${conf}" "${NEW_DIR}/${conf}"
